+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_params.py
@@ -53,6 +53,7 @@ def load_arguments(self, _):
             c.argument('years', type=int, default=None, arg_group='Credential')
             c.argument('create_cert', action='store_true', arg_group='Credential')
             c.argument('keyvault', arg_group='Credential')
+            c.argument('append', action='store_true', help='Append the new credential instead of overwriting.')
 
     with self.argument_context('ad') as c:
         c.argument('display_name', help='object\'s display name or its prefix')
@@ -99,6 +100,7 @@ def load_arguments(self, _):
         c.argument('include_inherited', action='store_true', help='include assignments applied on parent scopes')
         c.argument('assignee', help='represent a user, group, or service principal. supported format: object id, user sign-in name, or service principal name')
         c.argument('ids', nargs='+', help='space separated role assignment ids')
+        c.argument('include_classic_administrators', arg_type=get_three_state_flag(), help='list default role assignments for subscription classic administrators, aka co-admins')
 
     with self.argument_context('role definition') as c:
         c.argument('role_definition_id', options_list=['--name', '-n'], help='the role definition name')