+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_util.py
@@ -84,6 +84,10 @@ def get_sql_server_azure_ad_administrators_operations(kwargs):
     return get_sql_management_client(kwargs).server_azure_ad_administrators
 
 
+def get_sql_server_connection_policies_operations(kwargs):
+    return get_sql_management_client(kwargs).server_connection_policies
+
+
 def get_sql_server_keys_operations(kwargs):
     return get_sql_management_client(kwargs).server_keys
 
@@ -94,6 +98,10 @@ def get_sql_servers_operations(kwargs):
 
 def get_sql_server_usages_operations(kwargs):
     return get_sql_management_client(kwargs).server_usages
+
+
+def get_sql_subscription_usages_operations(kwargs):
+    return get_sql_management_client(kwargs).subscription_usages
 
 
 def get_sql_virtual_network_rules_operations(kwargs):