+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+2.0.20
+++++++
+* `acs create`: emit out an actionable error if provisioning application failed for lack of permissions
+
 
 2.0.19
 ++++++